+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 BUG FIXES:
 * resource/hcloud_ssh_key: Fix panic when we update labels in SSH keys
+* resource/hcloud_server_network: Fix alias ips ignored on creation of server network
 * resource/hcloud_server: Use first assigned `ipv6_address` as value instead of the network address. **Attention: This can be a breaking change**
 
 NOTES:
-* This release uses Terraform Plugin SDK v1.6.0.
+* This release uses Terraform Plugin SDK v1.8.0.
 
 ## 1.15.0 (November 11, 2019)